requirements.txt
alembic==1.9.4
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsons==1.6.3
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
nodeenv==1.7.0
packaging==23.0
prisma==0.8.1
prompt-toolkit==3.0.37
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.5.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
tomlkit==0.11.6
typing_extensions==4.5.0
typish==1.9.3
vine==5.0.0
wcwidth==0.2.6
webargs==8.2.0
Werkzeug==2.2.3
flask-swagger-ui==3.20.9
